Host CRUD is working

+++ app/host_routes.py
@@ -3,7 +3,7 @@ from app import db
 # from app.models.board import Board
 from .models.host import Host
 
-host_bp = Blueprint("/hosts", __name__, url_prefix="/")
+host_bp = Blueprint("/hosts", __name__, url_prefix="/hosts")
 
 #create a new host
 @host_bp.route("", methods=["POST"], strict_slashes=False)
@@ -12,10 +12,14 @@ def create_a_host():
     if ("Host name" not in request_body or "Introduction" not in request_body or "Phone number" not in request_body or "Address" not in request_body):
         return jsonify({"details": "Failed to create a host profile"}), 400
     
-    new_host = Host(hostname= request_body["Host name"],  introduction= request_body["Introduction"], address= request_body["Address"], phonenum= request_body["Phone number"])
+    new_host = Host(host_full_name=request_body["Host name"], 
+                    host_introduction=request_body["Introduction"], 
+                    host_address=request_body["Address"], 
+                    host_phone =request_body["Phone number"])
+    
     db.session.add(new_host)
     db.session.commit()
-    return jsonify({"Success": f"Host {new_host.hostname} is created"}), 201
+    return jsonify({"Success": f"Host {new_host.host_full_name} is created"}), 201
 
 #get a host profile
 @host_bp.route("/<host_id>", methods=["GET"])
@@ -24,7 +28,7 @@ def get_a_host_profile(host_id):
     if host== None:
         return jsonify({"Error": "Host is not found"}, 404)
     else:
-        return make_response(host.get_host_info(), 200)
+        return make_response(host.get_host_info_id(), 200)
     
 #get all the hosts in that location
 @host_bp.route("", methods=["GET"])
@@ -39,12 +43,18 @@ def get_all_hosts_profiles():
 
 
 #delete a host profile
-@host_bp.route("/<user_id>", methods=["DELETE"])
+@host_bp.route("/<host_id>", methods=["DELETE"])
 def delete_a_host_profile(host_id):
+    
     host = Host.query.get(host_id)
+    
+    if host == None:
+        return jsonify({"details" : f"No host with ID number {host_id} found "},404)
+    
     db.session.delete(host)
     db.session.commit() #how can we transfer this to a seperate column of deleted hosts
-    return jsonify({"Success": "Host is deleted"}, 200)
+    
+    return jsonify({"Success": "Host is deleted"}), 200
 
 #deactivate the host temporarily
